feat: api-videos seed update
